requirements.txt
streamlit
mysql-connector-python
python-dotenv
google.generativeai
pandas